speedy: msm_fb: fix overlays Glacier and Speedy overlays do not update the screen with just MSMFB_OVERLAY_PLAY - a manual update must be done.
HTC chose to do this in android::Overlay::queueBuffer, but I
have opted to place this support in the kernel as it seemed
much cleaner this way.

NOTE: kanged the fix from zinx. later on merge to main tree i think the config option name should be changed. its no longer device specific to the Glacier and will most likely be needed for other devices. i removed the Glacier config check so now on config of MDP it will ask. BE WARNED ;)

Conflicts:

	arch/arm/mach-msm/Kconfig